create shared_examples_for_exportable_by_version
